+++ heuristics.h
@@ -78,4 +78,6 @@ void free_list(PM_weight_pair_node* list_to_free);
 float power_consumption (float **utilization, int **H, int h_size, int *working_pms);
 void economical_revenue (VM_linked_list** vm_list, VM_linked_list** VM_list_derived, long *revenue, long *qos, int *living_vms, int *living_derived_vms);
 float wasted_resources (float **utilization, float **resources_requested, int **H, int h_size, float *wasted_resources_obj);
-float calculates_weighted_sum(float power, long total_revenue, float wasted_resources_ratio, long total_qos);
+float calculates_weighted_sum(float power, double total_revenue, float wasted_resources_ratio, double total_qos);
+
+
